backend/requirements.txt
asgiref==3.7.2
async-generator==1.10
attrs==22.2.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.2.1
django-extensions==3.2.3
django-rest-framework==0.1.0
djangorestframework==3.14.0
exceptiongroup==1.1.1
h11==0.14.0
idna==3.4
mysql-connector-python==8.0.33
mysqlclient==2.1.1
numpy==1.24.3
outcome==1.2.0
pandas==2.0.2
protobuf==3.20.3
psycopg2-binary==2.9.6
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
requests==2.28.2
selenium==4.8.3
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
sqlparse==0.4.4
trio==0.22.0
trio-websocket==0.10.2
typing_extensions==4.6.3
tzdata==2023.3
urllib3==1.26.15
wsproto==1.2.0
